Add fix for google services battery drain bug. by Alcolawl@xda
http://forum.xda-developers.com/google-nexus-5/general/script-google-play-services-battery-t3059585/post59563859

The Problem: Without going into much detail SystemUpdateService is basically Google's OTA update service,
and when it detects that your phone need and update while running a Custom ROM,
it fails and re-runs this check over and over again. It's this wakelock that drains your battery.

The Solution: Enabling the service. You read that correctly. For whatever reason,
enabling the service actually prevents the wakelock from occurring.
I, along with a few testers,
have been using our phones for hours now after having run the script and the wakelock and battery drain is completely gone.